+++ tests/testthat/setup-utils-lapply.R
@@ -2,8 +2,6 @@ future_is_installed <- ("future.apply" %in% installed.packages())
 
 os_is_windows <- (toupper(Sys.info()["sysname"]) == "WINDOWS")
 
-bettermc_is_installed <- ("bettermc" %in% installed.packages())
-
 parallel_is_installed <- ("parallel" %in% installed.packages())
 
 M <- 5